refactor: moved instructions nop and unreachable to control dir
