+++ scanner/instruction_parser.hpp
@@ -1,6 +1,14 @@
 #ifndef ANTIWASM_INSTRUCTION_PARSER_HPP
 #define ANTIWASM_INSTRUCTION_PARSER_HPP
 
+#include "control/br.hpp"
+#include "control/br_if.hpp"
+#include "control/br_table.hpp"
+#include "control/call.hpp"
+#include "control/call_indirect.hpp"
+#include "control/nop.hpp"
+#include "control/return.hpp"
+#include "control/unreachable.hpp"
 #include "instruction.hpp"
 #include "lexems.hpp"
 #include "memory/data_drop.hpp"
@@ -11,7 +19,6 @@
 #include "memory/memory_init.hpp"
 #include "memory/memory_size.hpp"
 #include "memory/store.hpp"
-#include "nop.hpp"
 #include "numeric/extension.hpp"
 #include "numeric/f32_const.hpp"
 #include "numeric/f64_const.hpp"
@@ -20,9 +27,6 @@
 #include "numeric/numericInstr.hpp"
 #include "numeric/saturating_truc.hpp"
 #include "numeric/type_conversion.hpp"
-#include <reference/ref_null.hpp>
-#include <reference/ref_is_null.hpp>
-#include <reference/ref_func.hpp>
 #include "table/elem_drop.hpp"
 #include "table/table_copy.hpp"
 #include "table/table_fill.hpp"
@@ -31,19 +35,15 @@
 #include "table/table_init.hpp"
 #include "table/table_set.hpp"
 #include "table/table_size.hpp"
-#include "unreachable.hpp"
 #include "variable/global_get.hpp"
 #include "variable/global_set.hpp"
 #include "variable/local_get.hpp"
 #include "variable/local_set.hpp"
 #include "variable/local_tee.hpp"
-#include "control/return.hpp"
-#include "control/br.hpp"
-#include "control/br_if.hpp"
-#include "control/br_table.hpp"
-#include "control/call.hpp"
-#include "control/call_indirect.hpp"
 #include <memory>
+#include <reference/ref_func.hpp>
+#include <reference/ref_is_null.hpp>
+#include <reference/ref_null.hpp>
 
 namespace antiwasm {
 std::unique_ptr<Instruction> parseInstruction(const uint8_t *instructionContent);
