+++ test/scanner/instruction_parser_test.cpp
@@ -5,7 +5,9 @@
 #include "control/br_table.cpp"
 #include "control/call.cpp"
 #include "control/call_indirect.cpp"
+#include "control/nop.cpp"
 #include "control/return.cpp"
+#include "control/unreachable.cpp"
 #include "instruction.cpp"
 #include "instruction_parser.cpp"
 #include "memory/data_drop.cpp"
@@ -16,7 +18,6 @@
 #include "memory/memory_init.cpp"
 #include "memory/memory_size.cpp"
 #include "memory/store.cpp"
-#include "nop.cpp"
 #include "numeric/extension.cpp"
 #include "numeric/f32_const.cpp"
 #include "numeric/f64_const.cpp"
@@ -33,7 +34,6 @@
 #include "table/table_init.cpp"
 #include "table/table_set.cpp"
 #include "table/table_size.cpp"
-#include "unreachable.cpp"
 #include "variable/global_get.cpp"
 #include "variable/global_set.cpp"
 #include "variable/local_get.cpp"
